+++ CHANGELOG.md
@@ -7,7 +7,8 @@ All notable changes to this project will be documented in this file.
 - Persistent volume for PostgreSQL in docker-compose
 
 ### Changed
-- Remove attribute `system` from BackendSource and add attribute `name` to ResourceRepresentation.
+- Remove attribute `system` from `BackendSource` and add attribute `name` to `ResourceRepresentation`.
+- Move attribute `sourceType` from `ResourceRepresentation` as `type` to `BackendSource`.
 
 ## [3.2.1] - 2020-11-05